Create product with Image

+++ MidasShopSolution.Data/Entites/ProductImage.cs
@@ -9,7 +9,6 @@ public class ProductImage
     public bool IsDefault { get; set; }
     public DateTime DateCreated { get; set; }
     public int SortOrder { get; set; }
-    public int FileSize { get; set; }
-    // public IFormFile ImageFile { get; set; }
+    public long FileSize { get; set; }
     public Product Product { get; set; }
 }
